// sl-billing-back/src/user/dto/create-user.dto.ts
import { IsEmail, IsNotEmpty, Matches } from 'class-validator';

export class CreateUserDto {
  @IsNotEmpty()
  firstName: string;

  @IsNotEmpty()
  lastName: string;

  @IsEmail()
  email: string;

  @IsNotEmpty()
  @Matches(/((?=.*\d)|(?=.*\W+))(?![.\n])(?=.*[A-Z])(?=.*[a-z]).*$/, {
    message: 'password is not a format',
  })
  password: string;

  @IsNotEmpty()
  role: string;

  avatar?: string;

  permission?: string[];

  confirmePassword?: string;

  teams?: [];
}